🚚 Break up large template module
