+++ lib/build.js
@@ -5,7 +5,7 @@ const commonjs = require("rollup-plugin-commonjs");
 const { terser } = require("rollup-plugin-terser");
 const { BUNDLE_FILENAME } = require('./constants.js');
 
-const template = require('./modules/template.js');
+const template = require('./modules/template/template.js');
 
 const build = async ({ input, output }) => {
