+++ lib/utils/fs.js
@@ -1,0 +1,10 @@
+const fs = require('fs');
+const { promisify } = require('util');
+
+const readFile = promisify(fs.readFile);
+const writeFile = promisify(fs.writeFile);
+
+module.exports = {
+    read: readFile,
+    write: writeFile,
+}
